requirements.txt
flask
requests
psycopg2-binary
redis
apscheduler
python-dotenv